Move common parser tests to `core:network` module
