+++ core/network/src/commonTest/kotlin/dev/sasikanth/rss/reader/core/network/CanaryTest.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2025 Sasikanth Miriyampalli
+ *
+ * Licensed under the GPL, Version 3.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.gnu.org/licenses/gpl-3.0.en.html
+ *
+ */
+
+package dev.sasikanth.rss.reader.core.network
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class CanaryTest {
+
+  @Test
+  fun canaryTest() {
+    assertTrue(true)
+  }
+}
